test(orchestration): split preparation coverage by owner
